moved lda logic into subdirectory of src
