+++ src/examples/NIPS_example.py
@@ -4,8 +4,8 @@ from time import time
 import numpy as np
 import pandas as pd
 
-from src.inference import get_top_n_words
-from src.sampler import LatentDirichletAllocation
+from src.lda.inference import get_top_n_words
+from src.lda.sampler import LatentDirichletAllocation
 
 if __name__ == "__main__":
     zf = zipfile.ZipFile("../Data/NIPS_1987-2015.csv.zip")
